requirements.txt
google-generativeai
pandas
openpyxl
requests
curl_cffi
python-dotenv
fastapi
uvicorn
supabase
python-multipart
python-jose[cryptography]
